LAMMPS: Also reconstruct FixQEqReax object when reconstructing PairReaxC object in lammpslib
The rebuild() function, introduced in ase/master@8700581, is called when
either the number of atoms in your system or the type of any of the
atoms has changed since the last evaluation (after calling
LAMMPSlib.initialise_lammps()).  If the number of atoms has changed, it
calls 'create_atoms' or 'destroy_atoms' and, in any case, updates the
species of all current atoms.  Within rebuild(), there's a special if
clause to see if reax/c is being used.  If so, it calls pair_style and
pair_coeff (first with a dummy lj/cut potential) in order to force the
current PairReaxC object to be destroyed and for a new one to be
created.  This is done because of the way reax allocates memory --- it
looks at the initial number of atoms in the system to makes its initial
allocations and then has some sort of (relatively fragile) monitoring
mechanism as the simulation proceeds to determine if it needs more.
However, what isn't being accounted for there is the 'fix qeq/reax' that
basically always accompanies 'pair_style reax/c'. When 'fix qeq/reax' is
called, a FixQEqReax object is created and, within its constructor, it
creates a reference to the current PairReaxC object. Thus, when
rebuild() forces a new PairReaxC object to be created, the reference to
the old PairReaxC object that's being destroyed still remains in the
FixQEqReax object and, when it attempts to reference this object, a
segfault occurs.