ENH: Add toy model user01-user20 species
The addition of these species is motivated by the KIM calculator
(ase.calculators.kim.kimmodel), which works with KIM Models which are
special in the sense that they are numerical models that represent a
general class of physical behavior and are not tied to any specific
atomic species.

Notes:
(1) The function ase.formula.parse2, which parses strings into the
individual species present, needed to be adjusted because our 'userXX'
species strings start with a lower-case letter and also because the
two-digit code following them does not indicate multiplicity.

(2) In what is done here, you *are* allowed to specify multiplicity by
placing one or more integers after the 'userXX' in the string.  For
instance, 'user013' is mapped to ['user01', 'user01', 'user01'].