LAMMPS: Set 'free_energy' in lammpslib to potential energy
This is an internal stopgap currently adopted in various classical
calculators solely to make sure that their energies are identified as
"force_consistent" (i.e. the forces are the numerical derivative of the
energy being reported) in optimizer or other output.  The term
"free_energy" is coming from a fictitious electronic temperature when
doing smearing to speed up convergence in ab initio calculations, but
it's not exposed in the classical calculators anyway. This mechanism
will probably change in the future.