LAMMPS: Name mass parameter consistently
In ase.calculators.lammpsrun.LAMMPS.set_missing_parameters, the
parameter 'masses', if omitted, if filled in based on the species
present. However, this is inconsistent with what is done in
ase.calculators.lammps.inputwriter.write_lammps_in, where the parameters
are checked for the key 'mass'. This means that if no 'mass' parameter
were specified when instantiating the calculator, the 'masses' parameter
set by set_missing_parameters() wouldn't actually get used to set any
masses.