DEV: Scaled convergence criteria in NEB
Add-on to the NEB class. Convergence criteria can be scaled to focus
optimization on saddle point and converge other interior states to
a higher fmax. Convergence scaling is currently implemented as a
displacement metric, where the fmax on each state depends upon the
position of the state in relation to the highest point in the band.
This can significantly reduce the number of force calls while giving
identical results at the saddle point. Only implemented in series
with dynamic_relaxation=True.