BUG: Ensure lammpsbc is called before set_cell
If any dimensions are shrink-wrapped and you try to change the cell size
using the LAMMPS change_box command, it will apply the shrink-wrapping
before it exists, negating any changes to the cell dimensions that you
made.  Therefore, we must scan for changes in the PBCs before we try to
change the cell dimensions.