BUG: Switch from get_positions(wrap) to atoms.wrap
Using wrap=True in the get_positions function is currently bugged in the
event that indefinite boundaries (cell length zero) are being used since
the inner product of the fractional positions with the cell is
performed.  Rather than wait for that to be fixed, we can just wrap the
atoms object in-place for now.