BUGFIX: CASTEP calculator detect conflicting basis keys and clobber
This is implemented with a general infrastructure as it is likely
other conflicting keyword sets will be added. As it stands the
implementation can only be used for sets of keywords within "param";
to manage conflicts between "param" and "cell" would require some kind
of communication between the setattr methods of the CastepParam and
CastepCell objects.