BUGFIX: Move Castep output cutoff energy to attribute from params
Read in the cutoff energy if provided at the end of a
calculation (this depends on the verbosity of the output).  However,
instead of setting it in the input parameters where it might clobber
BASIS_PRECISION, store it in calc._cut_off_energy in the same fashion
as timing information. This way it is available for e.g. analysis of
the on-the-fly pseudopotential generation, but does not cause
surprises in restarted calclations.