ENH: Allow Castep band structure to be specified with ASE BandPath
For now this is implemented as simply as possible, by using the kpt
list of BandPath to set bs_kpoint_list. Setting the band path with
set_kpts will reset BS_KPOINT(S)_XXX keywords but not KPOINT(S)_MP_XXX
keywords, and vice-versa. It is possible to set both at once using the
dictionary syntax. BS_KPOINT_SPACING will be ignored as it applies to
the unused BS_KPOINT_PATH keyword.

Note that at this point the Castep calculator lacks the capability
to *read* band structure information once computed.