BUGFIX: Vasp2 calculator restarting from files without ase-sort.dat
Response to problem in merge request !1274

The existing logic to build a fresh ase sort relies on
calculator.atoms, which has not been set yet when performing a
restart. As the next step is always to read CONTCAR, it should be
harmless to use CONTCAR here if no ase-sort was found.