BUG: change CASTEP output file 'regular end' for Mac compatibility
CASTEP does not check memory usage when built on Mac OSX as the
necessary system call is different. As a result, the Castep Calculator
incorrectly presumes that the calculation did not terminate properly.

Changing to 'Finalisation time =' based on a grep of the CASTEP
source code; this string only appears in one place (apart from test
data.)