ENH: Simplify compare_atoms handling of charges, magmoms
There are currently two quantities that may or may not be in an Atoms
object's `arrays` attribute: `initial_charges` and `initial_magmoms`.
The associated 'get' functions return an array of zeros for these if
they're not in `arrays`.  Accordingly, for our checks in compare_atoms,
we can simply assume that the system has changed if these exist in one
of the Atoms objects and not the other.