BUG: Fix update_kim_coords() function
In order to avoid a ton of explicit numpy array casts when initializing
some of its attributes, NeighborList has an override of __setattr__ that
creates a new numpy array and assigns it to the relevant attribute.
This interferes with the application of in-place addition (+=) to these
attributes in the sense that += will create a *new* object and assign it
to the attribute.  This is catastrophic in the case of the `coords`
attribute, since its exact address is registered with the KIM API and
must not be changed.  Calling update_kim_coords() was using += with
self.coords...

One option here is to always call update_compute_args after
update_kim_coords (as is currently done when a neighborlist update is
made), which will register the new address with the KIM API.  However,
since we only need to do in-place addition here, I decided it was
simplest to just use np.add with the 'out' kwarg set to the original
array (self.coords).