BUG: Add kim to list of special names
The function get_calculator_class() in calculator.py is failing when
called for the kim calculator because the appropriate attribute name is
"KIM" (upper case).  This is actually a function, but it always returns
a calculator, meaning that it behaves as though it were a class
definition for a calculator.