FIX: Update lammpsrun for lammpsdata format
lammpsrun uses NamedTemporaryFile to write a data file that's passed to
LAMMPS.  Since NamedTemporaryFile defaults to opening in binary mode and the
'lammps-data' format specification is '1F', we need to explicitly
request text mode.  Also note that technically LAMMPS data files must be
ASCII-encoded (although nothing actually attempts to write non-ASCII
characters to it and it's unlikely anything ever will...).