ENH: Provide wrapping option for LAMMPSlib calculator
This option allows one to disable the wrapping of the positions in the
Atoms object along periodic directions, which is sometimes useful for
tracking the path taken by individual atoms during minimization or
dynamics.  If the position of an atom in the Atoms object is more than
one box length outside of the cell along a periodic dimension, it will
be lost (but LAMMPS will error out in this case).

See also https://gitlab.com/ase/ase/issues/461