TST: use less strict tolerance for lammpslib tests
The tolerance for the lammpslib test seems too strict.
When tested against LAMMPS stable_12Dec2018, lammpslib_simple.py
failes with a relative difference of 4e-10, while the tolerance
is set to 1e-10.

Also, different assertions about accuraries are all changed to
np.testing.assert_allclose for consistancy.

The tolerances are new set to rtol=1e-7 for all lammpslib tests,
as is the default value of np.testing.assert_allclose.