Fix erroneous test failure
This test required a maximum energy deviation of 1e-10 after 10
molecular dynamics steps, which is probably too tight, and which
caused a failure on my system (which observed a difference of
about 4e-10). This threshold is far too tight, since we are just
making sure that LAMMPS is running, not that it is giving consistent
energies to its past behavior. The threshold has been loosened to 1e-8.