API: Remove lammpslib in-place atom wrapping
Reverting ab0f1399c7 in favor of removing in-place wrapping of the Atoms
object positions altogether.  Not only does in-place wrapping interfere
with optimizers, but it may generally confuse users and it breaks the
encapsulation of the calculator.

Since ab0f1399c7 (merged to master ~1 week ago) introduced the `wrap`
option, this change does break backward compatibility but I think it's
probably ok.