TST: XDATCAR test - make position-wrapping explicit
Explicitly set the wrap argument in get_scaled_positions. While this
is the default behaviour, it is a somewhat controversial choice. We
should preserve the intent to wrap in this case. so that trouble is
avoided when/if get_scaled_positions() is deprecated and the code
needs to be updated.