ENH: ensure atoms are wrapped by lammps inputwriter
If a LAMMPS data file (in the form of a string) is not passed to
ase.calculators.lammps.inputwriter.write_lammps_in, it will put a
'create_atoms' command into the input file for every atom in the Atoms
object.  This commit ensures that atoms outside of the cell along
periodic dimensions are wrapped inside of the box in this case.

Note that if a data file is passed to write_lammps_in, it just writes a
corresponding 'read_data' line into the input file, and 'read_data' in
LAMMPS always remaps the atoms.

Fixes: #514