RawDOSData: sample_grid padding, Gaussian function and much pytest
The sample_grid method is modified for RawDOSData to make use of the
broadening width and automatically pad the data range, avoiding a
sharp cutoff of the broadened data.

Pytest is used to check the values of single-peak broadening against
the usual peak-height and half-maximum solutions. The Gaussian
implementation was tweaked with some extra factors of two to
accomodate these standard definitions.

Plotting is tested by inspecting the Axes object.