DOSData: mark abstract methods and simplify arguments
Rather than annoy mypy and repeat code by allowing for future
versions of the sample() method with different arguments, keep things
simple and fix the broadening arguments to "width" and
"smearing". If we want to do something more complicated than Gaussian
or Lorentzian broadening, it probably shouldn't be happening here anyway.

Python has provided a way to explicitly mark base classes and abstract
methods, so we can use this for a bit of extra safety. In the test
framework we can "defuse" the restrictions by definining a minimal
inheriting class and testing _that_.