don't check if basisfile exists, and minor fix
If the basisfile doesn't exist, then either we will fail to read it in
(in case literal inclusion was requested), or Gaussian will fail to run
(in case we are just providing a path to Gaussian). In both cases, it is
better if the failure happens at the actual source of the problem.
Consider, for example, the scenario where I am creating an input file
asynchronously from running Gaussian. I don't necessarily want the check
to happen at input file creation time.

Also, make sure to invalidate previously-parsed results (energies,
forces, and dipole moment) even if we haven't encountered a molecular
geometry yet, because any results parsed before a molecular geometry has
been found probably refer to a geometry we failed to parse, or one that
was never printed in the output file.