unit conv fix, and don't parse imprecise dipole
Gaussian prints the dipole in (at least) two places, and with different
units. In one place, it prints with relatively high precision in atomic
units (e-Bohr), and in another, it prints with quite low precision in
Debye. Previously, the high-precision dipole units were not being
convertd correctly. Also, remove parsing of the low-precision dipole
because 1) it normally comes after the high-precision dipole (thus
overwriting it), and 2) it screws up trajectory parsing when a frequency
calculation is requested following a geometry optimization.