Gaussian: more dipole parsing fixes
So, Gaussian internally rotates the molecule as you provided it into a
"standard orientation", whatever that means. A lot of the calculations
are done in this orientation, except for the dipole moment it would
seem. However, sometimes Gaussian will print the dipole "rotated" into
the standard orientation. Previously, we parsed this dipole just like
any other dipole. Now we skip that dipole, because it probably isn't
correct given our understanding of the molecular geometry.