Selection and summing methods for DOSCollection
sum_all() and sum_by() provide means for condensing a
DOSCollection to simpler collection. select() and select_not() provide
means for splitting a DOSCollection into relevant groups. __add__()
provides a facility to join DOSCollections together. In combination
these tools should allow users to construct the PDOS plot of their
dreams.

total() is almost an alias to sum_all(), but also sets the label. It
should make code that adds a Total line to a PDOS a bit more clean and
"functional-looking" despite the ugly need to modify state in the info
field. Perhaps it is unnecessary cruft here and should be dropped?