DOSCollection: placating MyPy while using single-dispatch
Mypy and single-dispatch don't play together nicely; the recommended
idiom for single-dispatch is to name your registered alternate
functions as '_' (the names won't be used anyway). But mypy sees
redundant definitions and gets angry. The best compromise seems to be
to give these functions some (unused) meaningful names.

https://github.com/python/mypy/issues/2904

The other tweak for Mypy's benefit is to add a sanity check that no
failed select() statements (returning None) can slip into sum_by() and
be asked to call a sum_all() method. This *should* be an impossible
scenario (because of how the query was generated) but, well, this is
software engineering and nothing is impossible...