Move DOSData/DOSCollection equality test to private _almost_equals
Requested by Ask in !1724 : don't support an "equality" comparison for
floating-point data collections as it leads to trouble further down
the line. To keep testing _fairly_ clean we retain an internal object
method.

Note that that test _on_ this method in DOSCollection is no longer
using pytest.parametrize; this wasn't working properly before. You
can't put fixtures into parameter lists with stock pytest, but this is
possible with the pytest-lazy-fixture plugin. We should wait and see
how widespread the problem is before adding more dependencies.