Move ase.dft.band_structure to ase.spectrum.band_structure
The ase.spectrum directory has been created as part of DOS refactoring
work. It is a more logical home for band_structure than ase.dft, as
band structure diagrams may be used for properties other than DFT
eigenvalues, such as phonon frequencies.

For now it is made possible to import from the old location, which
will raise a deprecation warning. Under the current CI setup raising
this warning will mark a test as 'failed', ensuring that internal
references are made to the new location.