TST: Generalise test for module deprecation warnings in test_imports
As the test_imports loop over all modules needs modifying to "forgive"
modules raising deprecation warnings, it might as well hold the logic
to _require_ deprecation warnings.

Additional modules/aliases that raise deprecations warnings on import
simply need to be added to the deprecated_modules set, which is
located next to other sets of exceptional modules.