Fix the weights of k-points for nscf calculation (#58)
PR for changing the weights generation from 1 for all the k-points to 1/N, where N is the number of k-points. This could be important when kcw calculation is built upon the nscf one where k-grid was provided manually.
At the moment the generation of the uniform grid for the NSCF calculation erroneously assign weights = 1 to all the k-points. This PR correctly assign 1/Nk where Nk is the total number of k-points (we assume no symmetries). This is relevant when the mesh for the NSCF calculation is explicitly given from input (and not via the automatic option). Needed in special case like e.g. the interface with EPW which requires a regular grid with k-point \in [0;1)