Rename package to avoid nameclashes (#65)
This PR renames the python source folder from `ase` to `ase_koopmans`, in order to avoind nameclashes in python environments where both this package and the original `ase` are installed.

It also addes the parsing of `nelec` for `pw.x` calculations and removes the explicit construction of `amass` for ph calculations as these are generated automatically.